implemented editable comments from profile page

+++ hooks/ProfileHooks.ts
@@ -1,5 +1,9 @@
-export async function DeleteGuide(id:string){
 
+
+export async function DeleteGuide(id:string){
+    if(!id){
+        throw new Error('Something went wrong.')
+    }
         const resp = await fetch("/api/guides/delete_guide", {
             method: "DELETE",
             headers: {
@@ -11,7 +15,9 @@ export async function DeleteGuide(id:string){
         return msg 
 }
 export async function DeleteComment(id:string){
-
+    if(!id){
+        throw new Error('Something went wrong.')
+    }
     const resp = await fetch("/api/comments/delete_comment", {
         method: "DELETE",
         headers: {
@@ -22,8 +28,20 @@ export async function DeleteComment(id:string){
     const msg = await resp.json()
     return msg 
 }
-export async function EditComment(){
+export async function EditComment(id:string, message:string){
+    if(!id || !message){
+        throw new Error('Something went wrong.')
+    }
 
+    const resp = await fetch("/api/comments/update_comment", {
+        method: "PATCH",
+        headers: {
+            "Content-Type":"application/json"
+        },
+        body: JSON.stringify({id, message})
+    })
+    const msg = await resp.json()
+    return msg 
 }
 export async function EditGuide(){
